fix(*): added the maximum file size constrant
