+++ app/lib/constants.ts
@@ -15,6 +15,10 @@ export const DEFAULT_PAGE_LIMIT = 10;
 export const ACCEPTED_DOCUMENT_TYPES =
   ".pdf,.doc,.docx,.txt,.jpg,.jpeg,.png,.gif,.bmp,.webp";
 
+/** Maximum document upload size in bytes (25 MB) */
+export const MAX_DOCUMENT_SIZE_MB = 25;
+export const MAX_DOCUMENT_SIZE_BYTES = MAX_DOCUMENT_SIZE_MB * 1024 * 1024;
+
 /** Custom event dispatched when background validation invalidates the in-memory cache */
 export const CACHE_INVALIDATED_EVENT = "kaapi:config-cache-invalidated";
