+++ app/lib/types/document.ts
@@ -1,0 +1,9 @@
+export interface Document {
+  id: string;
+  fname: string;
+  object_store_url: string;
+  signed_url?: string;
+  file_size?: number;
+  inserted_at?: string;
+  updated_at?: string;
+}
